+++ src/main/java/net/geforcemods/securitycraft/blockentities/SecurityCameraBlockEntity.java
@@ -43,9 +43,10 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 	private double cameraRotation = 0.0D;
 	private double oCameraRotation = 0.0D;
 	private boolean addToRotation = SecurityCraft.RANDOM.nextBoolean();
-	private boolean down = false, downSet = false;
+	private boolean down = false, initialized = false;
 	private int playersViewing = 0;
 	private boolean shutDown = false;
+	private float initialXRotation, initialYRotation;
 	private DoubleOption rotationSpeedOption = new DoubleOption(this::getBlockPos, "rotationSpeed", 0.018D, 0.01D, 0.025D, 0.001D, true);
 	private BooleanOption shouldRotateOption = new BooleanOption("shouldRotate", true);
 	private DoubleOption customRotationOption = new DoubleOption(this::getBlockPos, "customRotation", getCameraRotation(), 1.55D, -1.55D, rotationSpeedOption.get(), true);
@@ -61,9 +62,14 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 
 	@Override
 	public void tick() {
-		if (!downSet) {
-			down = getBlockState().getValue(SecurityCameraBlock.FACING) == Direction.DOWN;
-			downSet = true;
+		if (!initialized) {
+			Direction facing = getBlockState().getValue(SecurityCameraBlock.FACING);
+
+			initialized = true;
+			down = facing == Direction.DOWN;
+
+			if (!isModuleEnabled(ModuleType.SMART))
+				setDefaultViewingDirection(facing);
 		}
 
 		oCameraRotation = getCameraRotation();
@@ -91,6 +97,8 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 		super.save(tag);
 		tag.putBoolean("shutDown", shutDown);
 		tag.put("lens", lens.createTag());
+		tag.putFloat("initial_x_rotation", initialXRotation);
+		tag.putFloat("initial_y_rotation", initialYRotation);
 		return tag;
 	}
 
@@ -99,6 +107,8 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 		super.load(state, tag);
 		shutDown = tag.getBoolean("shutDown");
 		lens.fromTag(tag.getList("lens", Constants.NBT.TAG_COMPOUND));
+		initialXRotation = tag.getFloat("initial_x_rotation");
+		initialYRotation = tag.getFloat("initial_y_rotation");
 	}
 
 	@Override
@@ -167,7 +177,7 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 	@Override
 	public ModuleType[] acceptedModules() {
 		return new ModuleType[] {
-				ModuleType.REDSTONE, ModuleType.ALLOWLIST
+				ModuleType.REDSTONE, ModuleType.ALLOWLIST, ModuleType.SMART
 		};
 	}
 
@@ -184,6 +194,8 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 
 		if (module == ModuleType.REDSTONE)
 			level.setBlockAndUpdate(worldPosition, getBlockState().setValue(SecurityCameraBlock.POWERED, false));
+		else if (module == ModuleType.SMART)
+			setDefaultViewingDirection(getBlockState().getValue(SecurityCameraBlock.FACING));
 	}
 
 	@Override
@@ -249,5 +261,39 @@ public class SecurityCameraBlockEntity extends CustomizableBlockEntity implement
 
 	public int getOpacity() {
 		return opacity.get();
+	}
+
+	public void setDefaultViewingDirection(Direction facing) {
+		float yRotation;
+
+		switch (facing) {
+			case NORTH:
+				yRotation = 180F;
+				break;
+			case WEST:
+				yRotation = 90F;
+				break;
+			case EAST:
+				yRotation = 270F;
+				break;
+			default:
+				yRotation = 0F;
+		}
+
+		setDefaultViewingDirection(down ? 75F : 30F, yRotation);
+	}
+
+	public void setDefaultViewingDirection(float initialXRotation, float initialYRotation) {
+		this.initialXRotation = initialXRotation;
+		this.initialYRotation = initialYRotation;
+		setChanged();
+	}
+
+	public float getInitialXRotation() {
+		return initialXRotation;
+	}
+
+	public float getInitialYRotation() {
+		return initialYRotation;
 	}
 }