+++ src/main/java/net/geforcemods/securitycraft/entity/camera/CameraController.java
@@ -5,6 +5,7 @@ import net.geforcemods.securitycraft.blocks.SecurityCameraBlock;
 import net.geforcemods.securitycraft.misc.KeyBindings;
 import net.geforcemods.securitycraft.misc.SCSounds;
 import net.geforcemods.securitycraft.network.server.DismountCamera;
+import net.geforcemods.securitycraft.network.server.SetDefaultCameraViewingDirection;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.GameSettings;
 import net.minecraft.client.Minecraft;
@@ -29,6 +30,7 @@ public class CameraController {
 	private static boolean wasDownPressed;
 	private static boolean wasLeftPressed;
 	private static boolean wasRightPressed;
+	private static int setDefaultViewingDirectionCooldown = 0;
 
 	private CameraController() {}
 
@@ -92,6 +94,11 @@ public class CameraController {
 
 				if (KeyBindings.cameraActivateNightVision.consumeClick())
 					giveNightVision(cam);
+
+				if (setDefaultViewingDirectionCooldown-- <= 0 && KeyBindings.setDefaultViewingDirection.consumeClick()) {
+					setDefaultViewingDirection(cam);
+					setDefaultViewingDirectionCooldown = 20;
+				}
 
 				//update other players with the head rotation
 				ClientPlayerEntity player = Minecraft.getInstance().player;
@@ -193,6 +200,10 @@ public class CameraController {
 	public static void giveNightVision(SecurityCamera cam) {
 		if (cam.toggleNightVisionCooldown == 0)
 			cam.toggleNightVisionFromClient();
+	}
+
+	public static void setDefaultViewingDirection(SecurityCamera cam) {
+		SecurityCraft.channel.sendToServer(new SetDefaultCameraViewingDirection(cam));
 	}
 
 	public static ClientChunkProvider.ChunkArray getCameraStorage() {